+++ p2p_wallet/Coordinators/NetworkCoordinator.swift
@@ -1,10 +1,3 @@
-//
-//  NetworkCoordinator.swift
-//  p2p_wallet
-//
-//  Created by Ivan on 31.08.2022.
-//
-
 import Combine
 import Foundation
 import UIKit
@@ -12,35 +5,21 @@ import UIKit
 final class NetworkCoordinator: Coordinator<Void> {
     private let navigationController: UINavigationController
 
-    private let transition = PanelTransition()
-
     init(navigationController: UINavigationController) {
         self.navigationController = navigationController
     }
 
     override func start() -> AnyPublisher<Void, Never> {
         let viewModel = NetworkViewModel()
-        let viewController = NetworkView(viewModel: viewModel).asViewController()
-        viewController.view.layer.cornerRadius = 16
-        viewController.transitioningDelegate = transition
+        let viewController = BottomSheetController(title: L10n.network, rootView: NetworkView(viewModel: viewModel))
         viewController.modalPresentationStyle = .custom
         navigationController.present(viewController, animated: true)
 
-        viewModel.dismiss
-            .sink(receiveValue: {
+        return Publishers.Merge(
+            viewModel.dismiss.handleEvents(receiveOutput: { _ in
                 viewController.dismiss(animated: true)
-            })
-            .store(in: &subscriptions)
-        transition.dimmClicked
-            .sink(receiveValue: {
-                viewController.dismiss(animated: true)
-            })
-            .store(in: &subscriptions)
-
-        let cancelSubject = PassthroughSubject<Void, Never>()
-        viewController.onClose = {
-            cancelSubject.send()
-        }
-        return cancelSubject.prefix(1).eraseToAnyPublisher()
+            }),
+            viewController.deallocatedPublisher()
+        ).prefix(1).eraseToAnyPublisher()
     }
 }
