+++ p2p_wallet/Coordinators/TabBar/ActionsCoordinator.swift
@@ -26,35 +26,38 @@ final class ActionsCoordinator: Coordinator<ActionsCoordinator.Result> {
             view.action.map { ActionsCoordinator.Result.action(type: $0) }
         )
 
-        view.action
-            .sink(receiveValue: { [unowned self] actionType in
-                switch actionType {
-                case .buy:
-                    break
-                case .receive:
-                    analyticsManager.log(event: AmplitudeEvent.actionButtonReceive)
-                    analyticsManager.log(event: AmplitudeEvent.mainScreenReceiveOpen)
-                    analyticsManager.log(event: AmplitudeEvent.receiveViewed(fromPage: "Main_Screen"))
-                case .swap:
-                    analyticsManager.log(event: AmplitudeEvent.actionButtonSwap)
-                    analyticsManager.log(event: AmplitudeEvent.mainScreenSwapOpen)
-                    analyticsManager.log(event: AmplitudeEvent.swapViewed(lastScreen: "Main_Screen"))
-                case .send:
-                    analyticsManager.log(event: AmplitudeEvent.actionButtonSend)
-                    analyticsManager.log(event: AmplitudeEvent.mainScreenSendOpen)
-                    analyticsManager.log(event: AmplitudeEvent.sendViewed(lastScreen: "Main_Screen"))
-                case .cashOut:
-                    analyticsManager.log(event: AmplitudeEvent.sellClicked(source: "Action_Panel"))
-                }
-            })
-            .store(in: &subscriptions)
-
         return Publishers.Merge(
-            result.handleEvents(receiveOutput: { [weak self] _ in
+            result.handleEvents(receiveOutput: { [weak self] action in
+                switch action {
+                case .action(let type):
+                    self?.logAnalytics(for: type)
+                default: break
+                }
                 self?.viewController.dismiss(animated: true)
             }),
             viewController.deallocatedPublisher().map { ActionsCoordinator.Result.cancel }
         ).prefix(1).eraseToAnyPublisher()
+    }
+
+    private func logAnalytics(for actionType: ActionsView.Action) {
+        switch actionType {
+        case .buy:
+            break
+        case .receive:
+            analyticsManager.log(event: AmplitudeEvent.actionButtonReceive)
+            analyticsManager.log(event: AmplitudeEvent.mainScreenReceiveOpen)
+            analyticsManager.log(event: AmplitudeEvent.receiveViewed(fromPage: "Main_Screen"))
+        case .swap:
+            analyticsManager.log(event: AmplitudeEvent.actionButtonSwap)
+            analyticsManager.log(event: AmplitudeEvent.mainScreenSwapOpen)
+            analyticsManager.log(event: AmplitudeEvent.swapViewed(lastScreen: "Main_Screen"))
+        case .send:
+            analyticsManager.log(event: AmplitudeEvent.actionButtonSend)
+            analyticsManager.log(event: AmplitudeEvent.mainScreenSendOpen)
+            analyticsManager.log(event: AmplitudeEvent.sendViewed(lastScreen: "Main_Screen"))
+        case .cashOut:
+            analyticsManager.log(event: AmplitudeEvent.sellClicked(source: "Action_Panel"))
+        }
     }
 }
